+++ src/components/pages/order/main/Basket/BasketProducts.js
@@ -1,15 +1,17 @@
 import React from "react"
 import { useContext } from "react"
 import styled from "styled-components"
-import OrderContext from "../../../../../context/OrderContext"
-import { findObjectById } from "../../../../../utils/array"
 import BasketCard from "./BasketCard"
+import { findObjectById } from "../../../../../utils/array"
 import { IMAGE_COMING_SOON } from "../../../../../enums/product"
+import OrderContext from "../../../../../context/OrderContext"
 
 export default function BasketProducts() {
-  const { basket, isModeAdmin, handleDeleteBasketProduct, menu } = useContext(OrderContext)
+  const { basket, isModeAdmin, handleDeleteBasketProduct, menu, handleProductSelected } =
+    useContext(OrderContext)
 
-  const handleOnDelete = (id) => {
+  const handleOnDelete = (event, id) => {
+    event.stopPropagation()
     handleDeleteBasketProduct(id)
   }
 
@@ -23,8 +25,9 @@ export default function BasketProducts() {
               {...menuProduct}
               imageSource={menuProduct.imageSource ? menuProduct.imageSource : IMAGE_COMING_SOON}
               quantity={basketProduct.quantity}
-              onDelete={() => handleOnDelete(basketProduct.id)}
+              onDelete={(event) => handleOnDelete(event, basketProduct.id)}
               isClickable={isModeAdmin}
+              onClick={isModeAdmin ? () => handleProductSelected(basketProduct.id) : null}
             />
           </div>
         )
