+++ src/components/pages/order/main/Basket/Total.js
@@ -1,14 +1,15 @@
 import styled from "styled-components"
 import Header from "../../../../reusable-ui/Header"
 import { theme } from "../../../../../theme"
-import { calculateSumToPay, formatPrice } from "../../../../../utils/maths"
+import { formatPrice } from "../../../../../utils/maths"
 import { useContext } from "react"
 import OrderContext from "../../../../../context/OrderContext"
+import { calculateSumToPay } from "./helper"
 
 export default function Total() {
   const { basket, menu } = useContext(OrderContext)
   const sumToPay = calculateSumToPay(basket, menu)
-  
+
   return (
     <Header>
       <TotalStyled>
