+++ src/components/pages/order/main/Basket/helper.js
@@ -1,0 +1,9 @@
+import { findObjectById } from "../../../../../utils/array"
+
+export const calculateSumToPay = (basket, menu) => {
+  return basket.reduce((total, basketProduct) => {
+    const menuProduct = findObjectById(basketProduct.id, menu)
+    total += menuProduct.price * basketProduct.quantity
+    return total
+  }, 0)
+}
