+++ src/context/OrderContext.js
@@ -3,16 +3,21 @@ import { createContext } from "react";
 export default createContext ({
     isModeAdmin: false,
     setIsModeAdmin: () => {},
+
     isTabSelected: "",
     setIsTabSelected: () => {},
+
     isCollapsed: false,
     setIsCollapsed: () => {},
+
     isAddSelected: true,
     setIsAddSelected: () => {},
     isEditSelected: false,
     setIsEditSelected: () => {},
+
     currentTabSelected: false,
     setCurrentTabSelected: () => {},
+
     menu: [],
     setMenu : () => {},
     handleAdd : () => {},
@@ -22,10 +27,10 @@ export default createContext ({
 
     newProduct: {},
     setNewProduct: () => {},
-    handleProductSelected: () => {},
 
     productSelected: {},
     setProductSelected: () => {},
+    handleProductSelected: () => {},
 
     titleEditRef: {},
