+++ src/utils/maths.js
@@ -1,29 +1,18 @@
-import { findObjectById } from "./array"
-
 export function formatPrice(priceToFormat) {
-    let price = priceToFormat
-  
-    // @TODO: perhaps change this to if(!price) return 0
-    if (!price) return "0,00 €"
-    price = replaceFrenchCommaWithDot(price)
-  
-    const formattedPrice = new Intl.NumberFormat("fr-FR", {
-      style: "currency",
-      currency: "EUR",
-    }).format(price)
-    return formattedPrice
-  }
-  
-  export function replaceFrenchCommaWithDot(price) {
-    if (typeof price === "string") price = parseFloat(price.replace(",", "."))
-    return price
-  }
+  let price = priceToFormat
 
-  export const calculateSumToPay = (basket, menu) => {
-    return basket.reduce((total, basketProduct) => {
-      const menuProduct = findObjectById(basketProduct.id, menu)
-      total += menuProduct.price * basketProduct.quantity
-      return total
-    }, 0)
-  }
-  
+  // @TODO: perhaps change this to if(!price) return 0
+  if (!price) return "0,00 €"
+  price = replaceFrenchCommaWithDot(price)
+
+  const formattedPrice = new Intl.NumberFormat("fr-FR", {
+    style: "currency",
+    currency: "EUR",
+  }).format(price)
+  return formattedPrice
+}
+
+export function replaceFrenchCommaWithDot(price) {
+  if (typeof price === "string") price = parseFloat(price.replace(",", "."))
+  return price
+}
